feat: Smuggle exception stack
This patch adds a method to `smuggle` which handles `ExceptionStack`s.
In the Julia REPL the `err` variable is implicitly defined and contains
the stack from the last thrown exception. Thus, for code evaluated in
the REPL directly (i.e. not through smuggling from neovim) this is very
useful since `smuggle(err)` can be used instead of having to manually
catch the error and backtrace.